signUp tab navigation 리펙토링
